move artifact resolution into a dedicated interface and implementation
