+++ plugin/src/main/kotlin/dev/aga/gradle/versioncatalogs/service/ArtifactResolver.kt
@@ -1,0 +1,8 @@
+package dev.aga.gradle.versioncatalogs.service
+
+import java.io.File
+import org.apache.maven.model.Dependency
+
+interface ArtifactResolver {
+    fun resolve(dep: Dependency, type: String = "pom"): File
+}
